TestFrameworks/pytest/pyproject.toml
[tool.poetry]
name = "pytest-test"
version = "0.1.0"
description = "pytest learning tutorial"
authors = ["Bulent Coskun <bulent.coskun@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pyleniumio = "^1.14.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
